+++ swagger.json
@@ -52,6 +52,48 @@
           "recipes"
         ],
         "operationId": "newRecipe",
+        "parameters": [
+          {
+            "type": "string",
+            "x-go-name": "Name",
+            "name": "name",
+            "in": "query"
+          },
+          {
+            "type": "array",
+            "items": {
+              "type": "string"
+            },
+            "x-go-name": "Tags",
+            "name": "tags",
+            "in": "query"
+          },
+          {
+            "type": "array",
+            "items": {
+              "type": "string"
+            },
+            "x-go-name": "Ingredients",
+            "name": "ingredients",
+            "in": "query"
+          },
+          {
+            "type": "array",
+            "items": {
+              "type": "string"
+            },
+            "x-go-name": "Instructions",
+            "name": "instructions",
+            "in": "query"
+          },
+          {
+            "type": "string",
+            "format": "date-time",
+            "x-go-name": "PublishedAt",
+            "name": "publishedAt",
+            "in": "query"
+          }
+        ],
         "responses": {
           "200": {
             "description": "Successful operation"
@@ -62,22 +104,20 @@
         }
       }
     },
-    "/recipes/search": {
+    "/recipes/{id}": {
       "get": {
-        "description": "Search recipes based on tags",
+        "description": "Get one recipe",
         "produces": [
           "application/json"
         ],
-        "tags": [
-          "recipes"
-        ],
-        "operationId": "findRecipe",
+        "operationId": "recipes",
         "parameters": [
           {
             "type": "string",
-            "description": "recipe tag",
-            "name": "tag",
-            "in": "query",
+            "x-go-name": "Name",
+            "description": "recipe ID",
+            "name": "id",
+            "in": "path",
             "required": true
           }
         ],
@@ -86,9 +126,7 @@
             "description": "Successful operation"
           }
         }
-      }
-    },
-    "/recipes/{id}": {
+      },
       "put": {
         "description": "Update an existing recipe",
         "produces": [
@@ -101,7 +139,7 @@
         "parameters": [
           {
             "type": "string",
-            "description": "ID of the recipe",
+            "description": "recipe ID",
             "name": "id",
             "in": "path",
             "required": true
@@ -131,7 +169,7 @@
         "parameters": [
           {
             "type": "string",
-            "description": "ID of the recipe",
+            "description": "recipe ID",
             "name": "id",
             "in": "path",
             "required": true
